// frontend/src/client/types.gen.ts
export type UserUpdateMe = {
    full_name?: (string | null);
    email?: (string | null);
};